Added more stuff

+++ SPICA/Utils/BitUtils.cs
@@ -6,18 +6,18 @@ namespace SPICA.Utils
 {
     class BitUtils
     {
-        public static byte GetBits(byte Value, int Start, int Count)
+        public static uint GetBits(uint Value, int Start, int Count)
         {
             Value >>= Start;
-            Value &= (byte)((1 << Count) - 1);
+            Value &= (uint)((1 << Count) - 1);
 
             return Value;
         }
 
-        public static byte SetBits(byte Value, byte Bits, int Start, int Count)
+        public static uint SetBits(uint Value, uint Bits, int Start, int Count)
         {
-            Value &= (byte)(~(((1 << Count) - 1) << Start));
-            Value |= (byte)(Bits << Start);
+            Value &= (uint)(~(((1 << Count) - 1) << Start));
+            Value |= Bits << Start;
 
             return Value;
         }
